% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagePreProcessing.R
\name{imagePreProcessing}
\alias{imagePreProcessing}
\title{imagePreProcessing}
\usage{
imagePreProcessing(zpFilepath)
}
\arguments{
\item{zpFilepath}{Filepath to the zooplankton image to be processed}
}
\value{
extractedImage The section of the image extracted by the algorithm,
                       or NULL if the image was corrupted
}
\description{
Preprocessing the passed image
}
\examples{
\dontrun{
zpFilepath <- 'Users/$(whoami)/Images/ZooplanktonImages/ostracod1.png'
ostracod1Image <- imagePreProcessing(zpFilepath)
}
}
