% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_image.R
\name{load_image}
\alias{load_image}
\title{load_image}
\usage{
load_image(filePath)
}
\arguments{
\item{filePath}{A string that contains the path to a .jpg, .jpeg, .png, or .tiff image to be loaded}
}
\value{
image An image (in greyscale) that corresponds to the filepath specified
}
\description{
Loads a Zooplankton image, so that they can be segmented and then classified
}
\examples{
filepath <- 'Users/$(whoami)/Images/ZooplanktonImages/ostracod1.png'
ostracod1Image <- load_image(filepath)
}
