% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_processedDataset.R
\name{unpack_processedDataset}
\alias{unpack_processedDataset}
\title{unpack_processedDataset}
\usage{
unpack_processedDataset(saveDir)
}
\arguments{
\item{saveDir}{(String) Directory to save the processed dataset to}
}
\description{
Unpacking the processed zooplankton dataset to a specific directory.
             The result will be a directory named "processedDataset"
             being put the specified save directory, containing all the image files
             from the processed dataset that comes with this package
}
\examples{
\dontrun{
saveDir <- 'Users/$(whoami)/Images/ZooplanktonImages/'
unpack_processedDataset(saveDir)
}
}
